Set up MVVM, made burger menu
