+++ .idea/workspace.xml
@@ -1,20 +1,44 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="AutoImportSettings">
-    <option name="autoReloadType" value="SELECTIVE" />
+    <option name="autoReloadType" value="NONE" />
   </component>
   <component name="ChangeListManager">
     <list default="true" id="83f1ba98-ac45-48c6-90e8-73adca6f18e6" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/lib/models/day.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/models/group.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/models/holiday.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/models/user.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/pubspec.yaml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/theme/app_theme.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/group_detail_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/holidays_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/login_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/navigation_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/shared_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/signup_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/viewmodels/today_viewmodel.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/views/auth/login_screen.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/views/auth/signup_screen.dart" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/lib/widgets/burger_menu.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/libraries/Dart_SDK.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/libraries/Dart_SDK.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/holidays_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/holidays_screen.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/aceup_clean.iml" beforeDir="false" afterPath="$PROJECT_DIR$/aceup_clean.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/biometric_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/views/auth/biometric_screen.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/group_detail_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/views/shared/group_detail_screen.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/holidays_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/views/holidays/holidays_screen.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/lib/main.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/main.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/lib/today_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/today_screen.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/pubspec.lock" beforeDir="false" afterPath="$PROJECT_DIR$/lib/pubspec.lock" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/shared_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/views/shared/shared_screen.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/lib/today_screen.dart" beforeDir="false" afterPath="$PROJECT_DIR$/lib/views/today/today_screen.dart" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="ClangdSettings">
+    <option name="formatViaClangd" value="false" />
   </component>
   <component name="FileEditorManager">
     <leaf>
@@ -32,6 +56,28 @@
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
+  <component name="GitHubPullRequestSearchHistory"><![CDATA[{
+  "lastFilter": {
+    "state": "OPEN",
+    "assignee": "Laudarias"
+  }
+}]]></component>
+  <component name="GithubPullRequestsUISettings"><![CDATA[{
+  "selectedUrlAndAccountId": {
+    "url": "https://github.com/ISIS3510-202520-T24/AceUp-Flutter",
+    "accountId": "edf57a3f-c1ba-4586-9365-7bbe92fb7bc8"
+  }
+}]]></component>
+  <component name="ProblemsViewState">
+    <option name="selectedTabId" value="CurrentFile" />
+  </component>
+  <component name="ProjectColorInfo"><![CDATA[{
+  "associatedIndex": 1
+}]]></component>
+  <component name="ProjectId" id="33QUT0vMK1UiOYtDjEWhIEbuRQg" />
+  <component name="ProjectLevelVcsManager">
+    <ConfirmationsSetting value="2" id="Add" />
+  </component>
   <component name="ProjectView">
     <navigator currentView="ProjectPane" proportions="" version="1" />
     <panes>
@@ -40,13 +86,46 @@
       </pane>
     </panes>
   </component>
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
   <component name="PropertiesComponent"><![CDATA[{
   "keyToString": {
+    "Flutter.main.dart.executor": "Run",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "RunOnceActivity.cidr.known.project.marker": "true",
+    "RunOnceActivity.readMode.enableVisualFormatting": "true",
+    "cf.first.check.clang-format": "false",
+    "cidr.known.project.marker": "true",
     "dart.analysis.tool.window.visible": "false",
-    "git-widget-placeholder": "7-8-create-today-and-holidays-views"
+    "git-widget-placeholder": "feature/25-implement-burger-menu",
+    "io.flutter.reload.alreadyRun": "true",
+    "project.structure.last.edited": "Modules",
+    "project.structure.proportion": "0.0",
+    "project.structure.side.proportion": "0.0",
+    "settings.editor.selected.configurable": "flutter.settings",
+    "show.migrate.to.gradle.popup": "false"
   }
 }]]></component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\LAURA\AceUp-Flutter\lib\views\shared" />
+      <recent name="C:\Users\LAURA\AceUp-Flutter\lib\views\auth" />
+      <recent name="C:\Users\LAURA\AceUp-Flutter\lib\views\today" />
+      <recent name="C:\Users\LAURA\AceUp-Flutter\lib\views\group_detail" />
+      <recent name="C:\Users\LAURA\AceUp-Flutter\lib\views\biometric" />
+    </key>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="83f1ba98-ac45-48c6-90e8-73adca6f18e6" name="Changes" comment="" />
+      <created>1759248804580</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1759248804580</updated>
+    </task>
     <servers />
   </component>
   <component name="ToolWindowManager">
