+++ lib/models/group.dart
@@ -1,0 +1,5 @@
+class Group {
+  final String name;
+
+  Group({required this.name});
+}
