+++ lib/models/holiday.dart
@@ -1,0 +1,6 @@
+class Holiday {
+  final String name;
+  final String dateRange;
+
+  Holiday({required this.name, required this.dateRange});
+}
