+++ lib/models/user.dart
@@ -1,0 +1,6 @@
+class User {
+  final String name;
+  final String email;
+
+  User({required this.name, required this.email});
+}
