+++ lib/viewmodels/login_viewmodel.dart
@@ -1,0 +1,6 @@
+class LoginViewModel {
+  bool login() {
+    // Replace with real auth logic
+    return true;
+  }
+}
