+++ lib/viewmodels/shared_viewmodel.dart
@@ -1,0 +1,8 @@
+import '../models/group.dart';
+
+class SharedViewModel {
+  final List<Group> groups = [
+    Group(name: "Study Group A"),
+    Group(name: "Team Project"),
+  ];
+}
