+++ lib/viewmodels/signup_viewmodel.dart
@@ -1,0 +1,5 @@
+class SignUpViewModel {
+  void signUp() {
+    // Replace with real signup logic
+  }
+}
