+++ lib/viewmodels/today_viewmodel.dart
@@ -1,0 +1,7 @@
+class TodayViewModel {
+  final List<String> tasks = [
+    "Math Homework",
+    "Group Meeting",
+    "Project Deadline",
+  ];
+}
