+++ lib/views/auth/login_screen.dart
@@ -1,0 +1,128 @@
+import 'package:flutter/material.dart';
+import '../../widgets/burger_menu.dart';
+import '../auth/signup_screen.dart';
+import '../auth/biometric_screen.dart';
+import '../../main.dart'; // for AppTheme
+
+class LoginScreen extends StatelessWidget {
+  const LoginScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+            child: Column(
+              children: [
+                const SizedBox(height: 40),
+                Image.asset('assets/logo.png', height: 120),
+                const SizedBox(height: 8),
+                RichText(
+                  textAlign: TextAlign.center,
+                  text: const TextSpan(
+                    style: TextStyle(
+                      fontSize: 36,
+                      fontWeight: FontWeight.w800,
+                      color: Color(0xFF0E2A47),
+                      letterSpacing: 0.5,
+                      height: 1.1,
+                    ),
+                    children: [
+                      TextSpan(text: 'Ace'),
+                      TextSpan(
+                        text: 'Up',
+                        style: TextStyle(fontWeight: FontWeight.w900),
+                      ),
+                    ],
+                  ),
+                ),
+                const SizedBox(height: 18),
+                const SizedBox(height: 24),
+                Align(
+                  alignment: Alignment.centerLeft,
+                  child: Text(
+                    'Welcome Back!',
+                    style: Theme.of(context)
+                        .textTheme
+                        .headlineSmall
+                        ?.copyWith(fontWeight: FontWeight.w700),
+                  ),
+                ),
+                const SizedBox(height: 12),
+                const TextField(
+                  decoration: InputDecoration(hintText: 'Email Address'),
+                ),
+                const SizedBox(height: 12),
+                TextField(
+                  obscureText: true,
+                  decoration: InputDecoration(
+                    hintText: 'Password',
+                    suffixIcon: IconButton(
+                      onPressed: null,
+                      icon: const Icon(Icons.visibility_off),
+                    ),
+                  ),
+                ),
+                const SizedBox(height: 8),
+                Align(
+                  alignment: Alignment.centerLeft,
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      TextButton(
+                        onPressed: () {},
+                        child: const Text('Forgot password?'),
+                      ),
+                      TextButton.icon(
+                        onPressed: () async {
+                          final result = await Navigator.pushNamed(context, '/biometric');
+                          if (result == true) {
+                            Navigator.pushReplacementNamed(context, '/today');
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              const SnackBar(content: Text('Biometric authentication OK')),
+                            );
+                          }
+                        },
+                        icon: const Icon(Icons.fingerprint),
+                        label: const Text('Login with Biometrics'),
+                      ),
+                    ],
+                  ),
+                ),
+                const SizedBox(height: 4),
+                SizedBox(
+                  width: double.infinity,
+                  height: 52,
+                  child: FilledButton(
+                    onPressed: () => Navigator.pushReplacementNamed(context, '/today'),
+                    style: FilledButton.styleFrom(
+                      backgroundColor: AppTheme.mint,
+                      foregroundColor: Colors.white,
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                    ),
+                    child: const Text('Login'),
+                  ),
+                ),
+                const SizedBox(height: 12),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    const Text('Not a member? '),
+                    TextButton(
+                      onPressed: () => Navigator.pushNamed(context, '/signup'),
+                      child: const Text('Register now'),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
